fix: fully resolve plugin paths when building using `From` (#562)
Introduce AbsPathRelativeToWorkspace() to Bazel interface.
Update bazel.New() to resolve provided paths an absolute path. This requires this function to return an error.
Introduce bazel.NoWorkspaceRoot() for commands that do not require a Bazel instance that has a workspace root (e.g., version).
Update error messages so that is easier to differentiate where an error occurred.