cli: fix custom plugin commands crashing (#1785)
Does this seem right? ~~I haven't verified this, but when~~ this was
added in
https://github.com/aspect-build/silo/commit/7473335ae9888851c8a35035009304a6daaa5d6a
I don't think any code is actually setting the value?

I made the same change in the `aspect-cli` repo, then changed the
`aspect-cli-plugin-template` `go_repository(aspect.build/cli)` to
`local_repository(../aspect-cli)` to reproduce and then test this fix.