fix: propogate error code from nested Aspect CLI when re-entering (#4130)
This is a large refactor to principally fix a subtle bug in
MaybeReenterAspect where the exit code from the re-entrant process is
not propagated to the exit code of the parent process.

TL;DR here is that the `bazelisk.Run` function is updated to not return
both an exitCode and err but rather just an error. If there is an
exitCode to propagate from Bazel then it is done so via the ExitError
struct at that point. This change unwinds many of the places that
ExitError is created from combining the excitCode and error.

---

### Type of change

- Bug fix (change which fixes an issue)

**For changes visible to end-users**

- Suggested release notes are provided below:

Aspect CLI will now correctly propagate the exit code of the nested
Aspect CLI process if a child Aspect CLI is spawned with the re-entry
logic.

### Test plan

- New test cases added