feat: add --bazel-version flag to Aspect CLI for a light-weight check of the underlying Bazel version (#4349)
Pre-factor to bazel version specific flags in Workflows.

Adding this flag snowballed into modifying the re-entrant logic since we
now need to ignore any bazelisk aspect CLI boostrap when the flag is
passed and just download bazel itself to call `bazel --version`. This
update to the re-entrant logic also fixes the flags-as-proto failure in
this
[thread](https://bazelbuild.slack.com/archives/C04281DTLH0/p1704997915936319),
which ended up being a necessary change to be able to add the
`--bazel-version` flag cleanly. These changes also affected the bats
tests for the Aspect CLI which were updated to use `.bazelversion` to
specify the bazel fetched bazel binary.

---

### Type of change

- New feature or functionality (change which adds functionality)

### Test plan

- New test cases added