fix: forward unrecognized flags to Bazel in query, cquery && aquery commands (#4659)
In cquery, aquery and query, we currently separate bazel flags from
other arguments since we need to be aware of args in those wrapped
commands for special handling. However, we error out on unrecognized
flags (flags that bazel help flags-as-proto doesn't report). This causes
problems when users pass flags such as
`--@io_bazel_rules_go//go/config:pure` which are defined by rule sets
and not known to `bazel help flags-as-proto`. We should instead be
forwarding these unrecognized flags to bazel to handle.

---

### Type of change

- Bug fix (change which fixes an issue)
- 
### Test plan

- Manual testing; please provide instructions so we can reproduce:

```
pro --aspect:lock_version cquery //cli/... --not-a-flag
```