feat(cli): support #aspect: gazelle directives (#5678)
A simple alias for now. Docs and things like `# gazelle:` references in
error messages and docs are not updated. Various tests have been changed
to use the alias.

`bazel-gazelle` in both the go.mod and `http_archive` are patched so
this applies to both the `gazelle_binary` and `bazel configure`.

---

### Changes are visible to end-users: no

### Test plan

- Covered by existing test cases
- New test cases added