fix(gazelle): only rename primary ts_project when name conflicts with npm_package (#5805)
The gazelle extension default `npm_package(name)` is the same as the
rules_js (v1) convention of `{dirname}`. However very often the "default
target" of a BUILD is the `ts_project` and not the `npm_package`, so
gazelle also uses `{dirname}` as the default `ts_project(name)`.

Then when generating both targets we have a naming conflict so we append
`_lib` to the `ts_project(name)`.

This renaming to avoid conflicts is the gazelle "default naming" and
should only be applied when using those defaults. Today it always
happens, we always rename the `ts_project` if an `npm_package` is also
being generated. If someone overrides the defaults then this appending
`_lib` default should no longer apply so that custom config is
predictable.

---

### Changes are visible to end-users: no

Users with custom naming conventions will no longer have the `_lib`
appended when both a `ts_project` and `npm_package` are generated in the
same BUILD.

### Test plan

- Covered by existing test cases
- New test cases added