fix(cli): treat package.json imports as optional dependencies (#7044)
There are too many edge cases that are causing issues atm. This should
have been optional or opt-in from the start.

This makes these imports "optional", so the feature is still enabled by
default and users upgrading the cli will get this new feature, but it
should no longer ever cause errors. We can maybe change it to error when
opted-in in the future and maybe change the default if we are more
certain it handles most of the edge cases.

---

### Changes are visible to end-users: yes

- Searched for relevant documentation and updated as needed: yes
- Breaking change (forces users to change their own code or config): yes
- Suggested release notes appear below: yes

Dependencies from package.json `main/types/exports` fields no longer
cause errors when `aspect configure` can not resolve the referenced
files.

### Test plan

- New test cases added