refactor(cli): upgrade configure directive suggestions to use aspect directives (#7025)
This is only error/suggestion messaging that the CLI outputs and does
not have any functional change, `# gazelle:` is still supported. Note
that we still refer to all the gazelle builtin directives, refer people
to gazelle docs etc, those docs will still have `# gazelle:` which could
cause some confusion? 🤷

Is this the direction we'd like to go? We are already requiring our
gazelle patches in order to use our gazelle plugins, support for `#
aspect:` is enabled with a 1-line patch. We could also do this only for
orion and not our existing js/kotlin extensions.

---

### Changes are visible to end-users: no

### Test plan

- Covered by existing test cases