perf(cli): avoid use of os.Stat for checking if file exists (#7173)
We should not be using `os.Stat` to check if files exist. It is very
expensive and gazelle has already read the full list of entries for the
directory.

The gazelle patch exposes that list. This is similar to
https://github.com/bazel-contrib/bazel-gazelle/pull/1737 which was done
at airtable for similar reasons.

On one major repo this reduces the time by almost 2s which is ~15% and
well worth the patch imo.

---

### Changes are visible to end-users: no

- Searched for relevant documentation and updated as needed: yes
- Breaking change (forces users to change their own code or config): no
- Suggested release notes appear below: yes

Reduce fs io of `aspect configure` during fs traversal.

### Test plan

- Covered by existing test cases