fix(CLI): Set BES completion timeout to 60 seconds (#7295)
When the `bazel lint` run finishes and we are waiting for the BES to complete, we had a 60 millisecond timeout. Increase this timeout to 60 seconds. 

Hit this timeout at a customer when we should not have

Customer thread: https://aspect-build.slack.com/archives/C03LLCZPA3D/p1728675490951849

---

### Changes are visible to end-users: yes

- Searched for relevant documentation and updated as needed: no
- Breaking change (forces users to change their own code or config): no
- Suggested release notes appear below: yes

Increase linting BES completion timeout to 60 seconds

### Test plan

- Covered by existing test cases