feat(cli): generate ts_config(visibility) attribute (#8838)
To continue limiting the (default) visibility of generated targets.

In scenarios where a base tsconfig is somewhere such as
`//tools/ts:tsconfig.json` this default will have to be overridden, but
I think that is fine for more complex repos with such structure.

---

### Changes are visible to end-users: yes

- Searched for relevant documentation and updated as needed: yes
- Breaking change (forces users to change their own code or config): no
- Suggested release notes appear below: yes

Generate `ts_config(visibility = [":__subpackages__"])` by default.

### Test plan

- Covered by existing test cases
- New test cases added