fix(cli): fully manage ts_project(isolated_typecheck) attribute in generated BUILDs (#9119)
I'm switching this to be fully managed by gazelle. Today we simply
assign a value but it is marked as `mergeable: false` so it will not
overwrite the attribute in the BUILD if it already exists. Now it is
`mergeable: true` so regenerating will correctly overwrite it. This has
been tested enough I'm confident we should be fully managing it now.

This attribute is still treated different then the "tsconfig attributes"
since it is not an actual tsconfig attribute. I've updated some comments
+ tests to make that clearer.

---

### Changes are visible to end-users: yes

- Searched for relevant documentation and updated as needed: yes
- Breaking change (forces users to change their own code or config): no
- Suggested release notes appear below: yes

Regenerate and overwrite `ts_project(isolated_typecheck)` each time
BUILDs are generated reflecting the tscofnig `isolatedDeclarations`
property.

### Test plan

- Covered by existing test cases
- New test cases added