refactor(cli): remove gazelle kotlin language from cli (#9178)
The gazelle kotlin `Language` can be loaded from the standalone
https://github.com/aspect-build/gazelle_kt

The only known users of this are using it via go.mod from the
`aspect-cli` repo, and they have forked it. They are already aware of
`gazelle_kt` and switching to it.

The tests were previously symlinked into `cli/pro/gazelle/tests/kotlin`,
this PR moves them there to continue testing the starlark implementation
of the kotlin BUILD-generation.

### Changes are visible to end-users: no

### Test plan

- Covered by existing test cases