refactor: add aspect-gazelle binary (#89)
`@aspect_gazelle_runner//bin/gazelle` is a binary designed to be
equivalent to a `gazelle_binary()` with additional aspect enhancements:
* a specific set of languages tested and verified by Aspect
* ability to disable those languages
* ability to add starlark extensions
* `--progress` for terminal progress
* support for being run with the aspect cli `--watch`
* opt-in to additional caching

The `@aspect_gazelle_runner//:def.bzl` provides a replacement for
`gazelle()` which uses the aspect binary and adds configuration of the
languages and starlark extensions as rule attributes.

### Changes are visible to end-users: yes

- Searched for relevant documentation and updated as needed: yes
- Breaking change (forces users to change their own code or config): no
- Suggested release notes appear below: yes

Provide an enhanced aspect-gazelle binary and `aspect_gazelle()` macro
as drop-in replacements for gazelle.

### Test plan

- Covered by existing test cases
- Manual testing; use the `aspect_gazelle()` target in place of
`gazelle()`