feat: support early exit for errors in gazelle languages run with aspect_gazelle_runner (#104)
This allows things like `--watch` mode to print the errors, stop the
gazelle execution (especially in the middle of a slow fs walk) and then
resume. Regular non-watch gazelle execution should remain unchanged and
do `os.Exit`.

The use of `context.Context` for early-exit is a proposed idea in
gazelle 2.0 and this this is designed to be easily migrated to that.

### Changes are visible to end-users: no

### Test plan

- Covered by existing test cases
- Manual testing; run the prebuilt gazelle (or cli `configure`) in watch
mode and observe `os.Exit` not being called